Extract BackendPool interface into three 3 separate interfaces
